範囲選択前のGUIのサーバーとローカルの差異を別クラスに分けた
