+++ Editor/EditorWindow/PlateauWindow/MainTabGUI/ImportGUIParts/ConfigGUIBeforeAreaSelectLocal.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 435bab4a84e4476fae601c8819703d4a
+timeCreated: 1702053620
