+++ Editor/EditorWindow/PlateauWindow/MainTabGUI/ImportGUIParts/ConfigGUIBeforeAreaSelectRemote.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: ca531a968fb14a618912a9f1658e6980
+timeCreated: 1702053747
