+++ Editor/EditorWindow/PlateauWindow/MainTabGUI/ImportGUIParts/CoordinateZonePopup.cs
@@ -1,5 +1,4 @@
-﻿using PLATEAU.CityImport.Config;
-using PLATEAU.Editor.EditorWindow.Common;
+﻿using PLATEAU.Editor.EditorWindow.Common;
 using PLATEAU.Geometries;
 using UnityEditor;
